Move schedule_hyperparams from wrappers.py to schedule.py.
