+++ optax/_src/schedule.py
@@ -20,12 +20,14 @@ instance, they may be used to anneal the learning rate used to update an agent's
 parameters or the exploration factor used to select actions.
 """
 
-from typing import Callable, Dict, Union, Optional
+from typing import Any, Callable, Dict, Union, Optional
 
 from absl import logging
 import chex
 import jax.numpy as jnp
 
+from optax._src import transform
+from optax._src import utils
 
 Schedule = Callable[[chex.Numeric], chex.Numeric]
 
@@ -345,3 +347,64 @@ def cosine_onecycle_schedule(
       peak_value / div_factor,
       {int(pct_start * transition_steps): div_factor,
        int(transition_steps): 1. / (div_factor * final_div_factor)})
+
+
+class ScheduleHyperParamsState(transform.OptState):
+  """Maintains inner transform state, hyperparameters, and step count."""
+  count: jnp.array  # shape=(), dtype=jnp.int32
+  hyperparams: Dict[str, chex.Numeric]
+  inner_state: transform.OptState
+
+
+def schedule_hyperparams(
+    inner_factory: Callable[..., transform.GradientTransformation],
+    **hyperparams: Union[Schedule, Any]
+) -> transform.GradientTransformation:
+  """Wrapper that schedules hyperparameters for the inner GradientTransform.
+
+  For example, to use SGD with an exponential decay for the learning rate,
+  piecewise constant schedule for momentum, and Nesterov momentum:
+
+  ```
+  scheduled_sgd = optax.schedule_hyperparams(
+      optax.sgd,
+      learning_rate=optax.exponential_decay(...),
+      momentum=optax.piecewise_constant_schedule(...),
+      nesterov=True)
+  ```
+
+  Args:
+    inner_factory: a function that returns the inner
+      `optax.GradientTransformation` given the hyperparameters.
+    **hyperparams: for each scheduled hyperparameter, there should be a function
+      that returns the hyperparameter value given a step count. For each
+      constant hyperparameter, there should be a corresponding value. The names
+      of these arguments must match the corresponding arguments for
+      `inner_factory`.
+
+  Returns:
+    An `optax.GradientTransformation`.
+  """
+
+  def schedule_fn(count):
+    return {k: (f(count) if callable(f) else f) for k, f in hyperparams.items()}
+
+  def init_fn(params):
+    count = jnp.zeros([], jnp.int32)
+    hparams = schedule_fn(count)
+    return ScheduleHyperParamsState(
+        count=count,
+        hyperparams=hparams,
+        inner_state=inner_factory(**hparams).init(params))
+
+  def update_fn(updates, state, params=None):
+    count_inc = utils.safe_int32_increment(state.count)
+    hparams = schedule_fn(count_inc)
+    updates, inner_state = inner_factory(**hparams).update(
+        updates, state.inner_state, params)
+    return updates, ScheduleHyperParamsState(
+        count=count_inc,
+        hyperparams=hparams,
+        inner_state=inner_state)
+
+  return transform.GradientTransformation(init_fn, update_fn)
